--- a/adadamp/_dist.py
+++ b/adadamp/_dist.py
@@ -26,13 +26,8 @@ Model = NewType("Model", torch.nn.Module)
 Grads = NewType("Grads", Dict[str, Union[torch.Tensor, float, int]])
 
 
-def _get_model_weights(model):
-    with torch.no_grad():
-        return sum(torch.abs(torch.sum(param)).item() for param in model.parameters())
-
-
-def _weight_sum(model_opt):
-    return _get_model_weights(model_opt[0])
+def _get_model_weights(model: nn.Module) -> float:
+    return sum(torch.abs(torch.sum(param)).item() for param in model.parameters())
 
 
 def _set_random_state(seed: int, device="cuda") -> bool:
@@ -95,7 +90,8 @@ def gradient(
     """
     # Workaround: Gradients should be cleared when entering this funciton,
     #     but at this moment this behavior is not occuring
-    model = deepcopy(model_opt[0])
+    #  model = deepcopy(model_opt[0])
+    model = model_opt[0]
 
     start = time()