**What does this PR implement?**
It performs the data transformations before any training begins. If a PyTorch dataset is passed to `.fit`, these transformations will be completed once every epoch (which might be slow). If the user wants to avoid that, they should use this code:

``` python
from adadamp import DaskBaseDamper
from torch.utils.data import transforms
from torchvision import datasets, transforms

net = DaskClassifier(module=Model, ...)

transform = transforms.Compose(
    [transforms.ToTensor(), transforms.Normalize((0.1307,), (0.3081,))]
)
ds = datasets.MNIST("./data", train=True, download=True, transform=transform)
ds2 = net.preprocess(df)
for epoch in range(10):
    net.partial_fit(ds2)
```