@stsievert 2021-02-08
This is a performance improvement: the transformations took 70% of time spent in `gradient`. I've looked at `decreasing-lr/exp--final.csv`, and here are the predicted timings:

|                | Avg. time for<br>`partial_fit`  | Equation |
| --------:     | :--- |:--- |
| Before this PR | 207.7ms | `t = partial_fit__time / n_updates` |
| After this PR  | ~62ms  | `t * (1 - 0.7)` |

<details>
<summary>Source equations</summary>

``` python
import pandas as pd
df = pd.read_csv("stats/decreasing-lr/exp--final.csv")
d = df[["partial_fit__time", "n_updates"]].copy()
running_avg = d["partial_fit__time"].cumsum() / df["n_updates"]
print(running_avg.iloc[-1])  # 0.207651...
```

I found that a call to gradient spent 70% of the time doing transformations by looking at `dask-report.html`, which says for the entire training, 810.76s were spent performing transformations and 1148.83s were spent in `gradient`, and `810 / 1148` is approximately `0.7055`.

</details>

I have a hard time believing this performance; 60ms is pretty fast. @JoeHolt can you spin up an EC2 instance and run a couple epochs (and save the CSV) to verify that the average time for a call to `gradient` is 60ms?