@stsievert 2021-02-12
I got the time for a round-trip model update down to 147ms. Here's the breakdown:

| Function | Time (ms) |
|:---|:------|
| `update_model` | 11 |
| `gradient`, deepcopy | 58 |
| `gradient` calculation | 78 |
| **Total** | 147 |

I think a comparison with the naive implementation should be generated. I think it'd be sufficient to take PyTorch's MNIST example and including this model/dataset. This would only be a timing comparison, so loss/optimizer/hyperparameters can be ignored.

I tried to compare with https://github.com/soumith/convnet-benchmarks/, but there are important size differences (image size, batch size, network size). It's in the same ballpark, but I can't tell if that's reasonable.